pyproject.toml
[project]
name = "ddong"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.12.0",
    "numpy>=2.3.5",
    "opencv-python>=4.11.0.86",
    "pillow>=12.0.0",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "transformers>=4.57.1",
]
